% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.scPDSI.file.R
\name{create.scPDSI.file}
\alias{create.scPDSI.file}
\title{Creates scPDSI output file.}
\usage{
create.scPDSI.file(
  scPDSI.file,
  f,
  ts,
  v.f.idx,
  variable.name.map,
  dim.size,
  dim.axes,
  scPDSI.dat,
  author.data
)
}
\arguments{
\item{scPDSI.file}{The filename to be used for the scPDSI file.}

\item{f}{The file(s) being used as sources for metadata.}

\item{ts}{The associated time data, as created by \code{nc.get.time.series}.}

\item{v.f.idx}{A mapping from variables to files, as created by \code{\link{get.var.file.idx}}.}

\item{variable.name.map}{A mapping from standardized names (tmax, tmin, etc) to NetCDF variable names.}

\item{dim.size}{Dimension sizes for the input.}

\item{dim.axes}{Dimension axes for the input.}

\item{scPDSI.dat}{scPDSI metadata}

\item{author.data}{A vector containing named elements describing the author; see \code{\link{create.indices.from.files}}.}
}
\value{
An object of class \code{ncdf4}.
}
\description{
Creates scPDSI output file.
}
\details{
This function creates a file suitable for outputting scPDSI to.
}
