% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{curry_in_subset_for_huglin}
\alias{curry_in_subset_for_huglin}
\title{A curry function used only for the Huglin Index
This function curries the cdx.funcs so that the current subset (cur_sub) is retrieved with the cdx function
It is placed inside compute.indices.for.stripe}
\usage{
curry_in_subset_for_huglin(cdx.funcs, cur_sub)
}
\description{
A curry function used only for the Huglin Index
This function curries the cdx.funcs so that the current subset (cur_sub) is retrieved with the cdx function
It is placed inside compute.indices.for.stripe
}
